+++ app_plus/src/main/java/org/gkisalatiga/plus/global/GlobalCompanion.kt
@@ -63,6 +63,7 @@ class GlobalCompanion : Application() {
         const val DEBUG_ENABLE_LOG_CAT_INIT = true
         const val DEBUG_ENABLE_LOG_CAT_LOCAL_STORAGE = true
         const val DEBUG_ENABLE_LOG_CAT_PDF = true
+        const val DEBUG_ENABLE_LOG_CAT_PERSISTENT_LOGGER = true
         const val DEBUG_ENABLE_LOG_CAT_PREFERENCES = true
         const val DEBUG_ENABLE_LOG_CAT_RAPID_TEST = true
         const val DEBUG_ENABLE_LOG_CAT_TEST = true