+++ app_plus/src/main/java/org/gkisalatiga/plus/services/InternalFileManager.kt
@@ -7,9 +7,14 @@
 package org.gkisalatiga.plus.services
 
 import android.content.Context
+import org.gkisalatiga.plus.lib.AppPreferences
 import org.gkisalatiga.plus.lib.LocalStorage
 import org.gkisalatiga.plus.lib.LocalStorageDataTypes
 import org.gkisalatiga.plus.lib.LocalStorageKeys
+import org.gkisalatiga.plus.lib.Logger
+import org.gkisalatiga.plus.lib.PersistentLogger
+import org.gkisalatiga.plus.lib.PreferenceKeys
+import java.io.File
 
 /**
  * Manages the app's internal file management.
@@ -28,12 +33,50 @@ class InternalFileManager (private val ctx: Context) {
     private val PDF_CONTENT_TUPLE_DELIMITER = ",,^,"
 
     /**
+     * Performs PDF cleaning-up.
+     * i.e., removal of PDF files not read after X period.
+     * This function does not delete a PDF file if it is marked as favorite.
+     */
+    fun doPdfCleanUp() {
+        LocalStorage(ctx).getAll().entries.filter { it.key.contains(LocalStorageKeys.LOCAL_KEY_IS_PDF_FILE_MARKED_AS_FAVORITE.name) }.forEach {
+            // Only clean-up PDF files that are not marked as favorite.
+            if (!(it.value as Boolean)) {
+                val url = LocalStorage(ctx).getDecomposeKey(it.key).second
+                val absolutePdfPath = LocalStorage(ctx).getLocalStorageValue(LocalStorageKeys.LOCAL_KEY_GET_CACHED_PDF_FILE_LOCATION, LocalStorageDataTypes.STRING, url) as String
+                val lastAccessMillis = LocalStorage(ctx).getLocalStorageValue(LocalStorageKeys.LOCAL_KEY_GET_PDF_LAST_DOWNLOAD_MILLIS, LocalStorageDataTypes.LONG, url) as Long
+                val prefKeepDaysOfCachedPdf = AppPreferences(ctx).getPreferenceValue(PreferenceKeys.PREF_KEY_KEEP_DAYS_OF_CACHED_PDF_FILES) as Long
+
+                // Debug only.
+                Logger.logTest({}, "lastAccessMillis: $lastAccessMillis, prefKeepDaysOfCachedPdf: $prefKeepDaysOfCachedPdf, currentTimeMillis: ${System.currentTimeMillis()}")
+
+                // Determines if we should keep this PDF file intact.
+                if (lastAccessMillis + prefKeepDaysOfCachedPdf <= System.currentTimeMillis()) {
+                    // Actually deleting the PDF file.
+                    File(absolutePdfPath).let { f -> if (f.exists()) f.delete() }
+
+                    // Remove any reference to the PDF file in the LocalStorage.
+                    LocalStorage(ctx).removeLocalStorageValue(LocalStorageKeys.LOCAL_KEY_GET_CACHED_PDF_FILE_LOCATION, url)
+                    LocalStorage(ctx).removeLocalStorageValue(LocalStorageKeys.LOCAL_KEY_GET_PDF_LAST_ACCESS_MILLIS, url)
+                    LocalStorage(ctx).removeLocalStorageValue(LocalStorageKeys.LOCAL_KEY_GET_PDF_LAST_DOWNLOAD_MILLIS, url)
+                    LocalStorage(ctx).removeLocalStorageValue(LocalStorageKeys.LOCAL_KEY_GET_PDF_METADATA, url)
+                    LocalStorage(ctx).removeLocalStorageValue(LocalStorageKeys.LOCAL_KEY_IS_PDF_FILE_DOWNLOADED, url)
+                    LocalStorage(ctx).removeLocalStorageValue(LocalStorageKeys.LOCAL_KEY_IS_PDF_FILE_MARKED_AS_FAVORITE, url)
+
+                    // Log the removal message persistently.
+                    PersistentLogger(ctx).write({}, "Removed PDF file due to clean-up of $prefKeepDaysOfCachedPdf millis: $url")
+                }
+            }
+        }
+    }
+
+    /**
      * (Primarily for use in ScreenPDFViewer.)
      * Enlists downloaded files and their local absolute path tuple in incremental manner.
      * The list will later be used to determine which PDF files to remove, to conserve space.
      * @param pdfAssociatedUrl the source URL associated with the downloaded PDF file.
      * @param pdfAbsolutePath the absolute path of the PDF file to be registered.
      */
+    @Deprecated("The PDF cleaning up now uses the serialization of all LocalStorage shared preferences pair reading.")
     fun enlistDownloadedFileForCleanUp(pdfAssociatedUrl: String, pdfAbsolutePath: String) {
         // Get the date of the enlisting of this PDF file.
         val currentDate = System.currentTimeMillis()